refactor(l2-remote): move the Request wire codec into RemoteL2LedgerCodecs
The Request Codec lived in the RemoteL2Ledger.Request companion; wire codecs
belong with the Response ones in RemoteL2LedgerCodecs. RemoteL2Ledger already
imports RemoteL2LedgerCodecs.given, so nothing else changes.
